+++ Tests/SourceKittenFrameworkTests/SourceKitTests.swift
@@ -244,17 +244,16 @@ class SourceKitTests: XCTestCase {
         XCTAssert(docsJSON.range(of: "error type") == nil)
         let jsonArray = try JSONSerialization.jsonObject(with: docsJSON.data(using: .utf8)!, options: []) as? NSArray
         XCTAssertNotNil(jsonArray, "JSON should be properly parsed")
-        let sourcekitd = "sourcekitd.framework/Versions/A/sourcekitd"
-        let sourcekitdInProc = "sourcekitdInProc.framework/Versions/A/sourcekitdInProc"
-        let modules: [(module: String, path: String, inProcPath: String?, linuxPath: String?)] = [
-            ("Clang_C", "libclang.dylib", nil, nil),
-            ("SourceKit", sourcekitd, sourcekitdInProc, "libsourcekitdInProc.so")
+        let sourcekitd = "sourcekitdInProc.framework/Versions/A/sourcekitdInProc"
+        let modules: [(module: String, macOSPath: String, linuxPath: String?)] = [
+            ("Clang_C", "libclang.dylib", nil),
+            ("SourceKit", sourcekitd, "libsourcekitdInProc.so")
         ]
-        for (module, path, inProcPath, linuxPath) in modules {
+        for (module, inProcPath, linuxPath) in modules {
             let wrapperPath = "\(projectRoot)/Source/SourceKittenFramework/library_wrapper_\(module).swift"
             let existingWrapper = try String(contentsOfFile: wrapperPath)
             let generatedWrapper = try libraryWrapperForModule(
-                module, loadPath: path, inProcLoadPath: inProcPath, linuxPath: linuxPath,
+                module, macOSPath: inProcPath, linuxPath: linuxPath,
                 compilerArguments: sourceKittenFrameworkModule.compilerArguments
             )
             XCTAssertEqual(existingWrapper, generatedWrapper)